cv_gui/requirements.txt
flask>=3.0.0
pyyaml>=6.0
gunicorn>=21.0.0
rendercv[full]>=2.0
